Initial commit. Adding a bunch of functionality right off the bat
This is the initial commit although there's a lot of existing code I'm
adding here... This is still very much in an experimental stage, and
there's a good chance much of these files/code will not be in the final
implementation. Just committing something to start off with in master so
there's something to roll back to.

This current code does compile and implements a simple neural network
for a creature brain although there isn't any real functionality yet