Refactor the way reproduction is done
Changed the way that creatures reproduce by basing it on a minimum
energy threshold instead of just age. This was reproduction is tied
to finding food directly.

This required some tweaks to the parameters and some slight cleanup/
refactoring of some of that code.

Also changed the behavior of the environment so that when creatures die,
they leave behind food instead of blank space. This increases the available
food in the environment and will eventually incentivize creatures
hunting each other.