Change movement actions to be relative to orientation (i.e forward/backwards)
Previously the 'brains' of the creatures had no way to know which
direction they should go in response to visual input as there is no
orientation input into the neural net and movement was absolute.

This is fixed by making movement relative (i.e. forward moves in the
direction that creature is facing)