Add ability for creatures to fight/kill each other
This adds another element to the simulation by allowing "Kill" to be
one of the actions a creature can take.

The creature gets immediate food for this and will slightly change it's
color to be more red to roughly indicate how violent it is. This should
allow other creatures to avoid violent ones (at least in theory)