Add feature to allow drawing new spaces
Add option to control panel that allows the user to modify the current
environment by drawing new wall/food spaces. This is intended to allow
the creation of new default environments that can be saved to JSON files
and loaded later.

It should allow for the creation of more interesting environments instead
of just randomly generated ones every time.

Also changed some of the default arguments to make things a little more
interesting