Translate project to "workspace" structure
This is a major change to the project structure that puts all
binaries in separate crates. This way we can more easily configure
them with different settings. For example, it was really difficult
to get the RGB matrix C library to compile only for the LED target.

This structure allows keeping the core library and different types
of visualization binaries separate.