Add creature color as a trait that can mutate
Add feature that allows creature color to be randomly changed just like
DNA when passed on to children. Instead of color indicating how violent
a creature is like before, color is inherited from the parent and will
be randomly mutated by a little bit each time according to the
`mutation_prob` parameter. This allows creatures to select for color
as an advantageous trait, and also has the side effect of allowing you to
kind of see which creatures are related (although that's not really
guaranteed)